+++ packages/common/integration/tests/pipes/brace.pipe.spec.ts
@@ -1,0 +1,21 @@
+import { BracePipe, BracePipeModule } from '@angular-ru/common/pipes';
+import { TestBed } from '@angular/core/testing';
+
+describe('brace pipe', () => {
+    let pipe: BracePipe;
+
+    beforeEach(() => {
+        TestBed.configureTestingModule({ imports: [BracePipeModule] }).compileComponents();
+        pipe = TestBed.inject(BracePipe);
+    });
+
+    it('should correct wrap input string with braces', () => {
+        const inputString: string = 'input';
+        expect(pipe.transform(inputString)).toEqual(`(${inputString})`);
+    });
+
+    it('should correct wrap input number with braces', () => {
+        const inputNumber: number = 42;
+        expect(pipe.transform(inputNumber)).toEqual(`(${inputNumber})`);
+    });
+});