+++ packages/tooltip/src/public_api.ts
@@ -1,3 +1,4 @@
+export { NgxTooltipTextInterceptor } from './ngx-tooltip-text.interceptor';
 export { TooltipDirective } from './tooltip.directive';
 export {
     TooltipContextValue,
@@ -9,6 +10,6 @@ export {
     TooltipValue
 } from './tooltip.interfaces';
 export { TooltipModule } from './tooltip.module';
-export { TOOLTIP_DEFAULT_OPTIONS } from './tooltip.properties';
+export { NGX_TOOLTIP_FALLBACK_PROVIDER, TOOLTIP_DEFAULT_OPTIONS } from './tooltip.properties';
 export { TOOLTIP_OPTIONS_TOKEN, TOOLTIP_TEXT_INTERCEPTOR_TOKEN } from './tooltip.tokens';
 export { TooltipDefaultTextInterceptor } from './tooltip-default-text.interceptor';